+++ Resources/Locale/en-US/xenoarchaeology/artifact-hints.ftl
@@ -41,3 +41,4 @@ artifact-trigger-hint-plasma = Gaseous plasma
 artifact-trigger-hint-land = Active deceleration
 artifact-trigger-hint-examine = Examination
 artifact-trigger-hint-medical = Therapeutic chemicals
+
