+++ Resources/Prototypes/Actions/types.yml
@@ -348,7 +348,7 @@
   name: Toggle Paramedic Siren
   description: Toggles the paramedic siren on and off.
   components:
-  - type: InstantAction
+  - type: InstantAction 
     icon:
       sprite: Clothing/OuterClothing/Hardsuits/paramed.rsi
       state: icon-siren
