+++ Resources/Prototypes/Entities/Mobs/NPCs/pets.yml
@@ -624,7 +624,6 @@
     - Hamster
     - VimPilot
     - ChefPilot
-    - MultitoolDefibrillatorTarget # Imp
   - type: FlavorProfile
     flavors:
     - meaty
