+++ Resources/Prototypes/Entities/Mobs/NPCs/regalrat.yml
@@ -210,6 +210,7 @@
     - map: [ "enum.DamageStateVisualLayers.BaseUnshaded"]
       state: eyes
       shader: unshaded
+
   - type: SpriteMovement
     movementLayers:
       movement:
@@ -294,7 +295,6 @@
     tags:
       - CannotSuicide
       - FootstepSound
-      - MultitoolDefibrillatorTarget # Imp
   - type: NoSlip
   - type: MobPrice
     price: 500 # rat wealth
