+++ Resources/Prototypes/Entities/Mobs/NPCs/space.yml
@@ -475,7 +475,6 @@
     - VimPilot
     - ChefPilot
     - Meat
-    - MultitoolDefibrillatorTarget # Imp
   - type: CombatMode
     combatToggleAction: ActionCombatModeToggleOff
   - type: Bloodstream
