+++ Resources/Prototypes/XenoArch/Effects/normal_effects.yml
@@ -60,7 +60,7 @@
 - type: artifactEffect
   id: EffectJunkSpawn
   targetDepth: 0
-  effectHint: artifact-effect-hint-retrieval
+  effectHint: artifact-effect-hint-retrieval # DeltaV - was hint-creation
   components:
   - type: SpawnArtifact
     maxSpawns: 10
@@ -106,7 +106,7 @@
 - type: artifactEffect
   id: EffectPointLight
   targetDepth: 0
-  effectHint: artifact-effect-hint-release
+  effectHint: artifact-effect-hint-release # DeltaV
   components:
   - type: PointLight
     radius: 8
@@ -149,7 +149,7 @@
   targetDepth: 0
   effectHint: artifact-effect-hint-environment
   components:
-  - type: ThrowArtifact
+    - type: ThrowArtifact
 
 - type: artifactEffect
   id: EffectChemicalPuddle
@@ -221,7 +221,7 @@
 - type: artifactEffect
   id: EffectInstrumentSpawn
   targetDepth: 1
-  effectHint: artifact-effect-hint-retrieval
+  effectHint: artifact-effect-hint-retrieval # DeltaV - was hint-creation
   components:
   - type: SpawnArtifact
     maxSpawns: 5
@@ -231,7 +231,7 @@
 - type: artifactEffect
   id: EffectMonkeySpawn
   targetDepth: 1
-  effectHint: artifact-effect-hint-retrieval
+  effectHint: artifact-effect-hint-retrieval # DeltaV - was hint-creation
   components:
   - type: SpawnArtifact
     spawns:
@@ -350,7 +350,7 @@
 - type: artifactEffect
   id: EffectFaunaSpawn
   targetDepth: 2
-  effectHint: artifact-effect-hint-retrieval
+  effectHint: artifact-effect-hint-retrieval # DeltaV - was hint-creation
   components:
   - type: SpawnArtifact
     maxSpawns: 5
@@ -564,7 +564,7 @@
 - type: artifactEffect
   id: EffectMaterialSpawn
   targetDepth: 3
-  effectHint: artifact-effect-hint-retrieval
+  effectHint: artifact-effect-hint-creation
   components:
   - type: SpawnArtifact
     maxSpawns: 5
