+++ Resources/Prototypes/XenoArch/Effects/utility_effects.yml
@@ -187,7 +187,7 @@
   id: EffectSentience
   targetDepth: 4 # DeltaV - was 3
   effectHint: artifact-effect-hint-sentience
-  effectProb: 0.5
+  effectProb: 0.5 # DeltaV
   permanentComponents:
   - type: GhostRole
     allowMovement: true
