+++ Resources/Prototypes/_DV/Entities/Mobs/NPCs/nukiemouse.yml
@@ -100,7 +100,6 @@
     - Meat
     - FootstepSound
     - Radio
-    - MultitoolDefibrillatorTarget # Imp
   - type: Puller
     needsHands: False
   - type: NoSlip
