+++ Resources/Prototypes/_DV/XenoArch/artifact_triggers.yml
@@ -37,3 +37,153 @@
   components:
   - type: ArtifactPressureTrigger
     maxPressureThreshold: 10000 # start a plasmafire or be clever with heaters
+
+- type: artifactTrigger
+  id: TriggerDrinkTheobromine
+  targetDepth: 4
+  triggerProb: 0.05
+  triggerHint: artifact-trigger-hint-drink-theobromine
+  components:
+  - type: Reactive
+    groups:
+      Acidic: [ Touch ]
+    reactions:
+    - reagents: [ Coffee, HotCocoa, IcedTea, Tea, ArnoldPalmer, PumpkinSpiceLatte, VodkaRedBool, XenoBasher, IrishBool, BudgetInsulsDrink, WatermelonWakeup, Rubberneck, EnergyDrink, Theobromine ]
+      methods: [ Touch ]
+      effects:
+      - !type:ActivateArtifact
+
+- type: artifactTrigger
+  id: TriggerDrinkGraveyard
+  targetDepth: 4
+  triggerProb: 0.05
+  triggerHint: artifact-trigger-hint-drink-graveyard
+  components:
+  - type: Reactive
+    groups:
+      Acidic: [ Touch ]
+    reactions:
+    - reagents: [ Graveyard ]
+      methods: [ Touch ]
+      effects:
+      - !type:ActivateArtifact
+
+- type: artifactTrigger
+  id: TriggerDrinkCircusJuice
+  targetDepth: 4
+  triggerProb: 0.05
+  triggerHint: artifact-trigger-hint-drink-circus-juice
+  components:
+  - type: Reactive
+    groups:
+      Acidic: [ Touch ]
+    reactions:
+    - reagents: [ CircusJuice ]
+      methods: [ Touch ]
+      effects:
+      - !type:ActivateArtifact
+
+- type: artifactTrigger
+  id: TriggerDrinkPinkDrink
+  targetDepth: 4
+  triggerProb: 0.05
+  triggerHint: artifact-trigger-hint-drink-pink-drink
+  components:
+  - type: Reactive
+    groups:
+      Acidic: [ Touch ]
+    reactions:
+    - reagents: [ PinkDrink ]
+      methods: [ Touch ]
+      effects:
+      - !type:ActivateArtifact
+
+- type: artifactTrigger
+  id: TriggerPumpkinSpiceLatte
+  targetDepth: 4
+  triggerProb: 0.05
+  triggerHint: artifact-trigger-hint-drink-pumpkin-spice-latte
+  components:
+  - type: Reactive
+    groups:
+      Acidic: [ Touch ]
+    reactions:
+    - reagents: [ PumpkinSpiceLatte ]
+      methods: [ Touch ]
+      effects:
+      - !type:ActivateArtifact
+
+- type: artifactTrigger
+  id: TriggerDrinkDoubleIceCream
+  targetDepth: 4
+  triggerProb: 0.05
+  triggerHint: artifact-trigger-hint-drink-double-ice-cream
+  components:
+  - type: Reactive
+    groups:
+      Acidic: [ Touch ]
+    reactions:
+    - reagents: [ DoubleIceCream ]
+      methods: [ Touch ]
+      effects:
+      - !type:ActivateArtifact
+
+- type: artifactTrigger
+  id: TriggerDrinkHealthCodeViolation
+  targetDepth: 4
+  triggerProb: 0.01
+  triggerHint: artifact-trigger-hint-drink-health-code-violation
+  components:
+  - type: Reactive
+    groups:
+      Acidic: [ Touch ]
+    reactions:
+    - reagents: [ HealthViolation ]
+      methods: [ Touch ]
+      effects:
+      - !type:ActivateArtifact
+
+- type: artifactTrigger
+  id: TriggerDrinkBrainBomb
+  targetDepth: 4
+  triggerProb: 0.01
+  triggerHint: artifact-trigger-hint-drink-brain-bomb
+  components:
+  - type: Reactive
+    groups:
+      Acidic: [ Touch ]
+    reactions:
+    - reagents: [ Brainbomb ]
+      methods: [ Touch ]
+      effects:
+      - !type:ActivateArtifact
+
+- type: artifactTrigger
+  id: TriggerDrinkAtomicBomb
+  targetDepth: 4
+  triggerProb: 0.01
+  triggerHint: artifact-trigger-hint-drink-atomic-bomb
+  components:
+  - type: Reactive
+    groups:
+      Acidic: [ Touch ]
+    reactions:
+    - reagents: [ AtomicBomb ]
+      methods: [ Touch ]
+      effects:
+      - !type:ActivateArtifact
+
+- type: artifactTrigger
+  id: TriggerDrinkMothamphetamine
+  targetDepth: 4
+  triggerProb: 0.01
+  triggerHint: artifact-trigger-hint-mothamphetamine
+  components:
+  - type: Reactive
+    groups:
+      Acidic: [ Touch ]
+    reactions:
+    - reagents: [ Mothamphetamine ]
+      methods: [ Touch ]
+      effects:
+      - !type:ActivateArtifact
