+++ Resources/Prototypes/_Impstation/tags.yml
@@ -3,6 +3,3 @@
 
 - type: Tag
   id: Holy
-
-- type: Tag
-  id: MultitoolDefibrillatorTarget
